+++ android/src/main/java/xyz/luan/audioplayers/Player.java
@@ -1,5 +1,6 @@
 package xyz.luan.audioplayers;
 
+import android.content.Context;
 abstract class Player {
 
     protected static boolean objectEquals(Object o1, Object o2) {
@@ -26,7 +27,7 @@ abstract class Player {
 
     abstract double getVolumeRight();
 
-    abstract void configAttributes(boolean respectSilence);
+    abstract void configAttributes(boolean respectSilence, boolean stayAwake, Context context);
 
     abstract void setReleaseMode(ReleaseMode releaseMode);